\name{SORS-package}
\alias{SORS-package}
\alias{SORS}
\docType{package}
\title{
  Smalleset overlap regions for copy number analysis
}
\description{
  The package finds out the smallest overlap regions (SORs) of copy number between samples. The functions to show heatmap and ideogram based on SORs are also provided in this package. 
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Gang Peng

Maintainer: Gang Peng <gang.peng@yale.edu>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ CNV, SOR }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
